requirements.txt
# Django & Core - Python 3.12 Compatible
Django==5.0.7
psycopg[binary,pool]==3.2.11

# Django Ninja - API Framework
django-ninja==1.3.0

# CORS
django-cors-headers==4.4.0

# Redis & Cache
redis==5.0.8
django-redis==5.4.0

# Environment Variables
python-dotenv==1.0.1

# Validation
pydantic==2.9.1
pydantic-settings==2.5.2
email-validator==2.2.0

# Authentication & Security
PyJWT==2.9.0
cryptography==43.0.1

# Utilities
python-slugify==8.0.4
pillow==10.4.0

# Production Server
gunicorn==23.0.0
whitenoise==6.7.0

# Python 3.12+ compatibility
setuptools>=70.0.0
wheel>=0.44.0
